Compiling, unit testing and Rollup fixes
Changes:

Updated Rollup to bundle files from ESNext to Javascript. Working on unit testing the engine to test running of the engine. Going to have to redesign and configure however. 

Issues:

* Still working on getting unit testing working
* Redesign of the engine needs to occur

Comments:

The engine needs to be redesigned. I'm debating between a static calling of the class or a variable to hold onto the object to check. Either way, also need to get the engine and each component to work separately for testing as well.