Added new engine files and unit testing
Changes:

Updated unit tests for the engine. Added DObject, GameObject, Component, and FlagComponent classes with test cases. Currently debugging and working on but would like to set a save point to continue working later. Working on GameObjects currently so that I test getting developer game files when deploying the engine to a new project. This will also be tested for when the developer compiles all the typescript files and deploys. This should be compiled through the engine when running.

Issues:

Unit test cases currently failing, working on solutions and implementations.

Comments:

Still working on the base engine to get it running and off the ground. Will As soon as some more base classes are in then I will work on bundling the package and running on Electron or Angular as a test.