GameWindow and Documents
Changes:

Added the GameWindow class. Should this have the direct link to the canvas or should the Render system 🤔? (I will ponder this). That said, this will be the container to the window (so maybe it should have all the udpate calls and such, but the renderer should touch it as well). Also added some documentation building to the `docs` folder now.

Issues:

The GameWindow class is going to take some working on to get right. Going to have to do some testing with how it will work with the render system.

Comments:

Add some prepublish scripts to build and deploy the documents. Maybe I can get a hook to a website for this? I don't know quite yet. Also going to have to get Travis-CI working sooner than later. Maybe those will be my tasks for tomorrow