Building files ready (#6)
* Building, unit testing, some base files

Changes:

- Setup base project files under `./src`
- Setup Rollup bundling to Javascript as ES and CommonJS file formats
- Using unit testing from Typescript files instead of the Javascript files; maybe look at this later
- gitignore and npmignore files setup

Issues:

- No unit tests yet
- No engine running yet
- Very touchy for the type of module loading I'm using; need to look into different module loading

Comments:
- Need to look into CommonJS vs ES vs AMD module loading
- For my project, should I look into building this just as a web app, or look into building this to an Electron application as well?
- I have a lot of devDependencies, is there a way to start cutting these down?

* Documentation building and compiling issues

Changes:

Added in documentation building with the 'npm run docs' command. Going to continue adding scripts to build the project, as I will eventually need to add in additional build steps for running the engine. Also updated the Rollup bundler to use Babel to compile from the projects target modules ESNext to CommonJS modules for consumption. This threw errors when running Mocha tests. I now have an environment variable being set by the 'unittest' script for the user.

Issues:

* None found at this time.

Comments:

None at this time.

* Compiling, unit testing and Rollup fixes

Changes:

Updated Rollup to bundle files from ESNext to Javascript. Working on unit testing the engine to test running of the engine. Going to have to redesign and configure however. 

Issues:

* Still working on getting unit testing working
* Redesign of the engine needs to occur

Comments:

The engine needs to be redesigned. I'm debating between a static calling of the class or a variable to hold onto the object to check. Either way, also need to get the engine and each component to work separately for testing as well.

* Added new engine files and unit testing

Changes:

Updated unit tests for the engine. Added DObject, GameObject, Component, and FlagComponent classes with test cases. Currently debugging and working on but would like to set a save point to continue working later. Working on GameObjects currently so that I test getting developer game files when deploying the engine to a new project. This will also be tested for when the developer compiles all the typescript files and deploys. This should be compiled through the engine when running.

Issues:

Unit test cases currently failing, working on solutions and implementations.

Comments:

Still working on the base engine to get it running and off the ground. Will As soon as some more base classes are in then I will work on bundling the package and running on Electron or Angular as a test.

* GameWindow and Documents

Changes:

Added the GameWindow class. Should this have the direct link to the canvas or should the Render system 🤔? (I will ponder this). That said, this will be the container to the window (so maybe it should have all the udpate calls and such, but the renderer should touch it as well). Also added some documentation building to the `docs` folder now.

Issues:

The GameWindow class is going to take some working on to get right. Going to have to do some testing with how it will work with the render system.

Comments:

Add some prepublish scripts to build and deploy the documents. Maybe I can get a hook to a website for this? I don't know quite yet. Also going to have to get Travis-CI working sooner than later. Maybe those will be my tasks for tomorrow

* Initial Readme

Changes:

Added an initial version of the readme. I've looked at other project's readmes, and this is one of the best templates I've found. Going to bulid off of it for my project. More to come and be added, but would like an initial file.

Issues:

See previous commit 7cc6c9a524eb1434c1d40fdaa55e55f2ec3b1c42.

Comments:

None

* Travis-CI script

Changes:

Added an initial Travis-CI script to run unit tests

Issues:

None

Comments:

None

* Specifying branches for Travis-CI and uploading documentation

Changes:

Fixed Travis file to run master and develop branches only. Also added documentation pushing and building to Github Pages from the docs folder.

Issues:

None

Comments:

None

* Set theme jekyll-theme-merlot

* Updated travis.yml

* Added Travis-CI button

None

* :confused: Had to update mocha

* Updated package.json and tslint



Changes:

[This should be an overall summary of changes to the project at least done in point-form. More detailed prose format is of course allowed. The point of the description is to provide detail of changes and why.]

Issues:

[Please provide a point-form description of issues. Then please submit a formal issue on the GitHub project. Issue guidelines are doucmented in the project Wiki.]

Comments:

[Other general comments you wish to add]

* I finally get it :fire:

Changes:

The developer will need to use my engine within their project - being browser, Electron or Angular based (working on just browser/electron first). 

Issues:

- When trying to test Electron.BrowserWindow within Mocha test, running into an error where the BrowserWindow constructor is undefined. For now I have the test not running at all.

Comments:

Developers first setup their own project THEN use my application within it (install or download it). Going to have to do some major revisions, but I think the best thing to do is make a diagram explaining this in a system architecture diagram. Going to have to do some thinking over the next week on how to do this, but also just need to sit down and work when I get back on it.

* Contributing file added

Changes:

Added the CONTRIBUTING.md file to the project tree in ready to submit the first projects pull request.

Issues:

none

Comments:

none