Electron Apps Working
Changes:

Tested and used the engine in an electron application. There was a major issue with the babel configuration and not properly transpiling the code to commonjs. The app now runs while creating a new Entity child called Player in the electron application and prints the ID to the console (this was in a test project and is not setup or shown here). This is good as it means the engine can be installed and used as is which is great. Now I need to work on the engine to see if it can get which platform it's being used on.

Issues:

* Working on electron but not Angular2 or direct HTML5 canvas

Comments:

None.