Github, core and tests
Changes:

Added a .github/ folder for all templates as I did not know that Github had this feature originally. Fixed up some tests, and added a new component. There needs to be some work done, but I have some ideas on how to approach next.

Issues:

None at this time.

Comments:

I will be travelling the next 3 weeks and will not be updating my code. Currently my proposal needs to be submitted again and that will happen after I am back from my trip. This commit is to make sure that all my work the past 2 weeks (not that I touched it much after the first day) is saved.