Puppeteer working, Divine global and Document aren't
Changes:

Puppeteer has been added to the Unit Tests to check UI and in browser variables of the engine. So far Divine is "undefined" but that is what I'm working to fix. New tests can be run with `npm test` and `npm run-script unit-test` as Istanbul now gives a coverage mapping as well.
Mostly build tools were added in this commit. There is lot's of debugging to come.

Issues:

* "Divine" must be a global variable in the browser scope
* GameWindow is still not being populated as "document" is undefined.

Comments:

[Other general comments you wish to add]