Puppeteer, Rollup and Modules
Changes:

Puppetter has been added to the project for Chrome testing. This will allow me to check and see if the "users'" application is running in the browser. However, as much as I would like to continue this right now, I am not having the best time with it. Therefore, I am moving on to another part of the project, the renderer. 
Rollup was previously spitting out the wrong formatted files, that has been fixed to UMD and ES modules (named accordingly).

Issues:

* Puppetter not working. After this commit, the tests will be turned off until I return to them.

Comments:

This has been a long time working at this, and needs to be put aside until I can find some help in the matter. This will be saved on a separate branch labelled "feature/puppeteer"