SceneManager, Scene and Engine
Changes:

Engine now has a SceneManager and loads in the Scene for the Engine. The Engine loads the basic scene for now, but it is HARDSET in Engine.ts (in the start method). SceneManager runs in tests, and also tests the Engine's startup of the scene. Scene tests still need to be written.
Helper functions have been written to read/write JSON data in files. This needs to be tested to see if the engine is pulling the files locally. This also needs to be tested to be running in a developers environment (Electron, HTML, etc). Helper functions to write the Scenes to JSON files should be tested. The idea is that the the scene will be written and read on start/stop of the Engines running to aid with compilation. Helper functions currently have no tests. These need to be added in when possible.
Scene file basic JSON file needs to be written. How are scenes going to be read/written?

Issues:

* Tests currently fail
    * File reading/writing has been hanging tests

Comments:

* Add todo items in files to the GitHub Project
* Add issues that are currently in the system to the Issues section
* TravisCI .yml file has been updated