WebGL, MessageSystem and Assets
Changes:

Message system has had a major overhaul; lot's of errors failing because of this reason. WebGL is beginning to be added in as I need to start getting work done on this. I'm loading this in with three.js and my package size has skyrocketed. Such as it needs to be however. Asset loading is next to begin testing. I need to see if I can get an asset loading for scene files.

Issues:

Current project issues:
* Asset loading needs to happen fast.
* Message tests need to be fixed up again...
Current relevant  GitHub issues:
* #28
* #11

Comments:

Time crunch is real.