Message system overhaul
Changes:

I have done a complete message system overhaul. As there was a lot of
circular dependency issues with the original system, I believe this may
be a soltuion. It still relies on an event based system, but it now also
has a priority cue order as well. To be included is a low, and urgent
cue, but this is to be implemented later - shouldn't be too bad as a
later task I have bigger fish to fry.
There is also an asset loader from the previous commit that will need to
be tested. This is the next priority item.
After asset loading testing is complete, triage is going to need to be
done on the rest of the project. Begin breaking down tasks and attacking
them head on.

Issues:

GitHub issues changed with this commit:
* #28: This commit will close issue #28 as the message system tests are
all working. The current failing tests are from the Engine itself.
* #11: An older issue that is more top level. I will reference this when
working on the physics engines' queue however, for now it can be closed.

Other issues:
* Currently the tests fail, going to have to continue debugging this
tomorrow.