Abstract System class
Changes:

*Major:*
Changing the System interface to an abstract class. Working on updating the other systems to this structure. Currently debating if systems should have their own cue or not. This will fall back to design of the engine. I need to sit and design from the current UML diagram of my system. Will work on this more tomorrow.

*Minor:*
Updated docs script. Now generates all the docs. The key was that I was using src/**/*.ts when I thought it would recursive open functions, but it does not. src/* fixes this.

Issues:

Comments:

This weekend was a blur and so will this month. Power through and do what I can. Sprint sprint sprint.