System, Render System, Physics System,
Changes:

System is now an abstract class (should have been in the first place)
    and will have all Systems as children. It just standardizes a few
    functions for the systems to use. Read the notes in the class for
    more information.

Render system is now using Three.js. This is going to cause some issues,
       and I'm going to branch here and give something a try before I
       continue. I need this commit as a stop point for now.

Physics system is a placeholder for the moment, but it "works".

All of these classes - and a vast majority more - of the engine needs to
be documented and tested to hit 100% and 70% (respectivly) for codecoverage and general
documentation.

Issues:

These will just need to get done, but I'm branching to try something
first.

Comments:

Here we go...