Adding some changes that I don't remember
I have not worked on this project for some time. This has been for
personal reasons, and I have needed the break to gather my thoughts. As
the title of this branch states, I will be giving this project a fresh
start; a new beginning from this point on. I have made many changes in
my workflow over the past couple of months, and I want to get this
project working right from here on. That means version updates, more
testing and beta testing as well as a trimmed down scope for what I
need. I'm going back to the basics: what does this project truly need
and what do I already have accomplished per my proposal.

For more information on this endeavour, please see issue #56.