revert(everything): Cleaning up the project
Cleaning up the project per issue #61. Going to start from here to follow my new GitFlow as talked
about in the [project wiki](https://github.com/srepollock/sunset-engine/wiki/Workflow). This will
also be the initial start - v0.0.1 - of the Sunset Engine. Looking forward to the work on this and
doing what I need to get this project completed to it's requirements (also in the
[wiki](https://github.com/srepollock/sunset-engine/wiki)).

re #61