Feature/message system (#68)
* Merge branch 'master' into hidden-master

* test(message system): Message system initial integration

Setup the mesasge system and tests to work in the engine. There are some test's failing, and I have
opened an issue to describe my changes (to which I'll be asking for help on StackOverflow tomorrow).
Almost done the message system, but for now I'm going to move onto the main engine loop and getting
an environment up and running in Docker I believe. This will be to setup development/production
variables as well as TravisCI testing on different platforms later on.

re #67

* test(messagesystem.spec.ts): Fixed process.stdout stream for TravisCI