feat(engine): Began creating the engine main loop
The main loop is currently failing in node due to the requestAnimationFrame not being part of the
core API (only in browser). Will look to extend a node loop process to run this otherwise. Currently
not using the message system either as it will require integration tests. Logging system was also
implemented, as I believe I left it out before. Will pickup tomorrow.

BREAKING CHANGE: Currently the unit tests do not run as the requestAnimationFrame does not exist in
Node