feat(Engine): Main loop
Implemented a majority of the main loop. Some debug testing is not passing currently so I will need
to update that tomorrow to get running. The unit tests are coming in handy as it appears it's not
running, but also I think I have to much in the tests at this time as well. All play testing at the
same time I suppose.