refactor(src/test): Refactored Engine code and Tests
We are now using Jest for our tests as it uses builtin Babel to comiple and run the code. Also, we
are targeting es2015 strictly, and will be using this as our base while writing ESNext code.
Confusing, but we are following ESNext -> es2015 -> bundled code for usage. Tests are failing as
there is still more to be written, but this took 2 days to refactor and run through, so it's an
ongoing process.

BREAKING CHANGE: Tests fail, but red/green testing is good