feat(scenemanager): Added saving methods
Setup the SceneManager and the DScene to have the saving methods as part of the class. The
implmentation will be done in the IOSystem and passed as a message through the MessageSystem (to be
tested later on). Did some minor fixes in the Engine, EngineArguments and a testing project in
Electron to get the game working better. Still have some errors with the window size, but this can
be fixed later on. Running into issues with the ThreeJS scene rendering, giving the warning:
Rendercount or primcount is 0. This looks to be an issue with the objects being added late to the
scene, or having an issue finding the scene. More debugging to come.