chore(rendersystem): Fixing up the render system
Working to change render system tests and render system in general. Going to remove ThreeJS (as I
can't get it working) and get my own WebGL context working. Going off of what I've learned in class
about WebGL. This will be better for my thought process and what I understand. Going to also change
up the tests to use Puppeteer or JSDom as it errors running Jest in the Node instance. This is going
to be some heavier work, so I wanted to get a checkpoint saved and continue with my work.