feat(matrix3/4): Created an inner Matrix3 / Matrix4 class
Created and tested Matrix3 and Matrix4 classes for usage in the engine. This will be used for WebGL
matrix convertions as gl-matrix could not be packaged and used correctly. Took extra time, but the
classes and their functions are all working correctly and can be used properly with WebGL now. Going
to continue creating a basic scene to render to the canvas now.