refactor(rendersystem): Not rendering
Worked to get the renderer working, and still nothing is happening. I think this may be more trouble
than it's worth, especially with the time frame that I have. I'm going to call it here to get
ThreeJS working again. It's should remove most of the complexity in this aspect.