feat: Rendering and Sprites
Worked on getting the sprite system implemented, but it began taking up too much time. Going to push
this off for now and continue working on getting movement and physics implemented first. Going to
take a bit of time, but I've got some ideas on the back burner for now. Also working on the report
along side work to get things done on time.