feat(rendersystem): Updating textures and tests
Worked on getting textures implemented into the engine. Cannot successfully get them loading in my
test electron project from the assets folder. Continuting to work on this over time. Working on
fixing the tests to get the working properly. There are a lot failing and out of date that need to
be udpated, so these are being addressed now. Continuing to work on them as the night progresses.