test(tests): Updated the system tests to reflect the changes for the message system
Tests have been updated with thresholds in the package.json file. Tests and the engine itself have
been updated to reflect the changes in the message system. These are now passing the basic unit
tests and should be good to go in the test project. Working on textures and the IOSystem next.