test: 💍 Added zones tests
Continuing to update the tests. This also has an integration test for
the ZoneManager as it requires the AssetManager and JsonAsset classes.
This required a bit of hacking together, but it works as intended with
some steps skipped (getting the JSON from an XMLHttpRequest, and waiting
for the message system to inform the ZoneMangaer that it had been loaded